requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-MySQL==1.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.9.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
